updated bar chart to look more like figma please i just want to go sleep
